server/requirements.txt
django==5.0.7
djangorestframework==3.15.1
django-environ==0.11.2
sqlparse==0.5.1
psycopg2==2.9.9
flake8==7.1.0
django-cors-headers==4.4.0
black==24.4.2
isort==5.13.2
gunicorn==22.0.0
